Update FivemPlayerlist.cs with an internal cache
The headshots are popped into another Dictionary that is periodically updated with the latest headshot of the players. This makes pressing `MULTIPLAYER_INFO` nice and fast!

I've also updated your `UpdateScale` function to not repeat the lines
```cs
row.textureString = textureCache[p.Handle] ?? "";
rows.Add(row);
```
It's not needed but, I think it's better as if you want to update it in future you can just change the one line instead of the two. 

With the textures being generated in another task, the original also didn't "await" anything so, I added a `Delay(0)` in there to keep everything happy.